+++ backend/src/auth/auth.controller.ts
@@ -1,6 +1,7 @@
-import { Controller, Request, Post } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { Controller, Post, Body, BadRequestException } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
+import { SigninReqDto, SignupReqDto } from './dto/req.dto';
+import { AuthService } from './auth.service';
 
 @ApiTags('Auth')
 @Controller('api/auth')
@@ -8,12 +9,39 @@ export class AuthController {
   constructor(private authService: AuthService) {}
 
   @Post('signup')
-  async signup() {
+  async signup(@Body() signupReqDto: SignupReqDto) {
     return this.authService.signup('email', 'password');
   }
 
   @Post('signin')
-  async signin(@Request() req) {
-    return this.authService.signin(req.user);
+  async signin(@Body() signinReqDto: SigninReqDto) {
+    return this.authService.signin({});
+  }
+
+  @Post('/oauth')
+  async createUserFromOAuth(@Body() requestBody: any) {
+    const { oauthProvider, oauthId, email, name, profilePictureUrl } =
+      requestBody;
+
+    if (!oauthProvider || !oauthId || !email || !name || !profilePictureUrl) {
+      throw new BadRequestException('Missing required fields.');
+    }
+
+    try {
+      const user = await this.authService.createOrUpdateUserFromOAuth(
+        oauthProvider,
+        oauthId,
+        email,
+        name,
+        profilePictureUrl,
+      );
+
+      return {
+        message: 'User created/updated successfully',
+        user,
+      };
+    } catch (error) {
+      throw new BadRequestException('Failed to create/update user.');
+    }
   }
 }
