+++ backend/src/auth/auth.module.ts
@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
+import { UserModule } from '@/user/user.module';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { PassportModule } from '@nestjs/passport';
-import { UserModule } from 'src/user/user.module';
 import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+import { User } from '@/entity/user.entity';
 
 @Module({
-  imports: [UserModule, PassportModule],
+  imports: [UserModule, PassportModule, TypeOrmModule.forFeature([User])],
   providers: [AuthService],
   controllers: [AuthController],
   exports: [AuthService],
